# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateNotificationSettingRequest:
    properties:
      body:
        type: string
      day_of_month:
        description: 0 = every day
        type: integer
      onesignal_player_id:
        type: string
      time_of_day:
        description: HH:MM
        type: string
      title:
        type: string
      type:
        type: string
    required:
    - body
    - onesignal_player_id
    - time_of_day
    - title
    - type
    type: object
  handlers.CreateScheduledFundRequest:
    properties:
      account_id:
        description: destination account
        type: string
      amount:
        type: number
      day_of_month:
        description: 1-31; 0 = last day
        type: integer
      description:
        type: string
      from_account_id:
        description: source for TRANSFER
        type: string
      schedule_type:
        description: TOP_UP | TRANSFER
        type: string
    required:
    - account_id
    - amount
    - schedule_type
    type: object
...
  handlers.RegisterDeviceRequest:
    properties:
      onesignal_player_id:
        type: string
    required:
    - onesignal_player_id
    type: object
...
  handlers.TestNotificationRequest:
    properties:
      body:
        type: string
      onesignal_player_id:
        type: string
      title:
        type: string
    required:
    - body
    - onesignal_player_id
    - title
    type: object
...
  handlers.UpdateNotificationSettingRequest:
    properties:
      body:
        type: string
      day_of_month:
        type: integer
      is_enabled:
        type: boolean
      onesignal_player_id:
        type: string
      time_of_day:
        type: string
      title:
        type: string
    type: object
  handlers.UpdateScheduledFundRequest:
    properties:
      amount:
        type: number
      day_of_month:
        type: integer
      description:
        type: string
      is_active:
        type: boolean
    type: object
...
  models.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@financeapp.com
    name: API Support
    url: https://github.com/yourusername/finance-tracking-app
  description: A comprehensive personal finance tracking application API with budget
    management, expense tracking, and analytics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Finance Tracking API
  version: "1.0"
paths:
...
  /notifications/register-device:
    post:
      consumes:
      - application/json
      parameters:
      - description: Device registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Register a device for push notifications (store player ID)
      tags:
      - notifications
  /notifications/settings:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get all notification settings
      tags:
      - notifications
    post:
      consumes:
      - application/json
      parameters:
      - description: Notification setting data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateNotificationSettingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create a push notification setting
      tags:
      - notifications
  /notifications/settings/{id}:
    delete:
      parameters:
      - description: Setting ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete a notification setting
      tags:
      - notifications
    get:
      parameters:
      - description: Setting ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get notification setting by ID
      tags:
      - notifications
    patch:
      consumes:
      - application/json
      parameters:
      - description: Setting ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateNotificationSettingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update a notification setting
      tags:
      - notifications
  /notifications/test:
    post:
      consumes:
      - application/json
      parameters:
      - description: Test notification data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.TestNotificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Send a test push notification
      tags:
      - notifications
...
  /scheduled-funds:
    get:
      parameters:
      - description: Filter active only
        in: query
        name: active_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get all scheduled funds
      tags:
      - scheduled-funds
    post:
      consumes:
      - application/json
      parameters:
      - description: Scheduled fund data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateScheduledFundRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create a recurring scheduled fund
      tags:
      - scheduled-funds
  /scheduled-funds/{id}:
    delete:
      parameters:
      - description: Scheduled Fund ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete a scheduled fund
      tags:
      - scheduled-funds
    get:
      parameters:
      - description: Scheduled Fund ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get a scheduled fund by ID
      tags:
      - scheduled-funds
    patch:
      consumes:
      - application/json
      parameters:
      - description: Scheduled Fund ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateScheduledFundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update a scheduled fund
      tags:
      - scheduled-funds
  /statistics/monthly:
    get:
      parameters:
      - description: 'Year (default: current year)'
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get monthly income/expense statistics for a year
      tags:
      - statistics
  /statistics/overview:
    get:
      parameters:
      - description: 'Month (1-12, default: current month)'
        in: query
        name: month
        type: integer
      - description: 'Year (default: current year)'
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get financial overview for a specific month
      tags:
      - statistics
...
schemes:
- http
- https
swagger: "2.0"